corrected a bad stash

+++ routes/item.js
@@ -33,21 +33,6 @@ const bcrypt = require("bcryptjs");
 });
 });
 
-// router.post("/createPost", (req,res) => {
-//     user.findOne({ username: req.body.username, email: req.body.email })
-//     .then(user => {
-//     if (!user) {
-//         return res.status(400).json({ message: "No such user exists" });
-//     } else {
-// const post = new item ({
-//     username: req.body.username,
-//     password: req.body.password,
-//     content: req.body.content
-//  });
-// // Save returns a promise
-// post.save().then(() => res.send('complete'));
-// });
-
 //@route GET content/read
 //@desc READ content
 //@access public
@@ -67,21 +52,53 @@ router.get("/getAll", (req,res) => {
 //@route PUT content/update
 //@desc UPDATE some content
 //@access public
-router.put("/update", (req,res) => {
-    item.updateOne({"username": req.body.username},
-    {$set: {"content": req.body.content}})
-    .then(() => 
-        res.send("updated")
-    );
+
+router.put("/updateLastPost", (req,res) => {
+
+    user.findOne({ username: req.body.username })
+.then(user => {
+if (!user) {
+    return res.status(400).json({ message: "No such user exists" });
+} else {
+    bcrypt.compare(req.body.password, user.password, function (err, result) {
+        if (result == true) {
+
+            item.updateOne({"username": req.body.username},
+            {$set: {"content": req.body.content}})
+            .then(() => 
+                res.send("updated")
+            );
+
+        } else {
+            res.send("Incorrect password");
+        }
+    });
+}
+});
 });
 
 //@route DELETE content/delete
 //@desc DELETE some content
 //@access public
-router.delete("/delete", (req,res) => {
-item.deleteOne({ "username": req.body.username })
-.then(() => 
-res.status(200).send("Successfully deleted"))
+
+router.delete("/deleteLastPost", (req,res) => {
+
+    user.findOne({ username: req.body.username })
+.then(user => {
+if (!user) {
+    return res.status(400).json({ message: "No such user exists" });
+} else {
+    bcrypt.compare(req.body.password, user.password, function (err, result) {
+        if (result == true) {
+            item.deleteOne({ "username": req.body.username })
+            .then(() => 
+            res.status(200).send("Successfully deleted"));
+        } else {
+            res.send("Incorrect Password");
+        }
+    });
+}
+});
 });
 
 module.exports = router;
